watchdog/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "watchdog-monitor"
version = "0.1.0"
description = "Async uptime & service monitor for Linux"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "pydantic>=1.10,<2.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.scripts]
watchdog = "main:main"

[tool.pytest.ini_options]
pythonpath = ["."]

